fix: shooting in autonomous

+++ include/robot/drivetrain.hpp
@@ -36,7 +36,7 @@ public:
    * @param max_rpm The maximum allowable RPM for the motor to run at while moving.
    * @param block Whether this function should wait for the robot to turn or exit immediately.
    */
-  void forwards(double distance, int32_t max_rpm = DRIVETRAIN_DEFAULT_RPM,
+  void forwards(double distance, int16_t max_rpm = DRIVETRAIN_DEFAULT_RPM,
                 uint16_t timeout_millis = DRIVETRAIN_DEFAULT_TIMEOUT);
 
   /**
@@ -45,7 +45,7 @@ public:
    * @param max_rpm The maximum allowable RPM for the motor to run at while moving.
    * @param block Whether this function should wait for the robot to turn or exit immediately.
    */
-  void backwards(double distance, int32_t max_rpm = DRIVETRAIN_DEFAULT_RPM,
+  void backwards(double distance, int16_t max_rpm = DRIVETRAIN_DEFAULT_RPM,
                  uint16_t timeout_millis = DRIVETRAIN_DEFAULT_TIMEOUT);
 
   /**
@@ -54,7 +54,7 @@ public:
    * @param max_rpm The maximum allowable RPM for the motor to run at while turning.
    * @param block Whether this function should wait for the robot to turn or exit immediately.
    */
-  void turn_right(double degrees, int32_t max_rpm = DRIVETRAIN_DEFAULT_RPM,
+  void turn_right(double degrees, int16_t max_rpm = DRIVETRAIN_DEFAULT_RPM,
                   uint16_t timeout_millis = DRIVETRAIN_DEFAULT_TIMEOUT);
 
   /**
@@ -63,7 +63,7 @@ public:
    * @param max_rpm The maximum allowable RPM for the motor to run at while turning.
    * @param block Whether this function should wait for the robot to turn or exit immediately.
    */
-  void turn_left(double degrees, int32_t max_rpm = DRIVETRAIN_DEFAULT_RPM,
+  void turn_left(double degrees, int16_t max_rpm = DRIVETRAIN_DEFAULT_RPM,
                  uint16_t timeout_millis = DRIVETRAIN_DEFAULT_TIMEOUT);
 
   void await_move(uint16_t timeout_millis = DRIVETRAIN_DEFAULT_TIMEOUT) const;
